Create package java.life.controller.
Package java.life.controller contains two classes,

Controller which controls and interacts with fxml components to display
prompts and updates throughout the game,

Player which gets invoked by Controller to initiate a game using
user-specified config. Player is the only ui-related class that
interacts with a Game object, and is reponsible for updating game grid
and side info pane.